+++ localvariable.cpp
@@ -259,7 +259,7 @@ CValue*	CLocalVariable::GetValuePtr(const aya::string_t &name)
  */
 int		CLocalVariable::GetNumber(int depth)
 {
-	if(depth < 0 || depth >= (int)stack.size()) {
+	if(depth < 0 || (size_t)depth >= stack.size()) {
 		return 0;
 	}
 	return stack[depth].substack.size();
